#1 Simplify file layout and merge functions in existing files
